refactor: add comments and add JSDoc's
