+++ src/scripts/add-effects.js
@@ -1,3 +1,8 @@
+/**
+ * Create a Audio and Image effects
+ * @param {object} main - Main object
+ */
+
 export default function (main) {
   main.sprites = new Image()
   main.sprites.src = '../src/assets/sprites.png'
