+++ src/scripts/create-background.js
