+++ src/scripts/create-canvas.js
