+++ src/scripts/create-floor.js
