+++ src/scripts/create-screen.js
@@ -1,3 +1,8 @@
+/**
+ * Create a screen.
+ * @param {object} main - Main object
+ */
+
 export default function (main) {
   main.changeScreen = newScreen => {
     main.screenActive = newScreen
@@ -5,6 +10,7 @@ export default function (main) {
     if (main.screenActive.started) {
       main.screenActive.started()
     }
+
     if (main.screenActive.over) {
       main.screenActive.over()
     }
