+++ src/scripts/draw-medals.js
