+++ src/scripts/makeFlappyPipeCollision.js
@@ -1,23 +1,23 @@
-import { global } from './display.js'
+import { global } from './create-display.js'
 
 /**
  * Function who create the collision between the pipe and the flappy bird.
- * @param {array} list
+ * @param {object} pipe - Pipe object.
  * @returns boolean
  */
 
-export default function flappyBirdCollision(list) {
+export default function flappyBirdCollision(pipe) {
   const flappyBird = global.flappyBird
 
   const flappyHead = flappyBird.canvasY
   const flappyFeet = flappyBird.canvasY + flappyBird.height
 
-  if ((flappyBird.canvasX + flappyBird.width)>= list.x) {
-    if (flappyHead <= list.pipeSky.y) {
+  if ((flappyBird.canvasX + flappyBird.width)>= pipe.x) {
+    if (flappyHead <= pipe.pipeSky.y) {
       return true
     }
 
-    if (flappyFeet >= list.pipeFloor.y) {
+    if (flappyFeet >= pipe.pipeFloor.y) {
       return true
     }
   }
